+++ assembler/src/include/omni/edges_position_handler.hpp
@@ -360,10 +360,9 @@ public:
 		if (path.size() > 0) {
 			vector<EdgePosition> res = (tmp_pos[path[0]]);
 
-			DEBUG(
-					this->g().int_id(path[0]) << "  "<< res.size() << " positions");
+			DEBUG(this->g().int_id(path[0]) << "  " << res.size() << " positions");
 
-			int len = this->g().length(path[0]);
+			int len = (int) this->g().length(path[0]);
 			for (size_t i = 1; i < path.size(); i++) {
 				DEBUG(this->g().int_id(path[i]) << "  "<< tmp_pos[path[i]].size() << " positions");
 				if (is_careful())
@@ -372,7 +371,7 @@ public:
 				else
 					res = GluePositionsLists(res, tmp_pos[path[i]],
 							max_single_gap_);
-				len += this->g().length(path[i]);
+				len += (int) this->g().length(path[i]);
 			}
 			if (res.size() > 0) {
 				if (is_careful()) {
